Rice codec is working. Cleanup up 'ricecodec' code. Copied sources to tsc 'src' folder.
Former-commit-id: 49d41ca1ca023acc7a7aaac369e1ecbe6dd8080b

+++ playground/ricecodec/main.c
@@ -32,7 +32,7 @@ int main(int argc, char* argv[])
 
     /* Compress or decompress. */
     unsigned char* out;
-    unsigned int out_sz = 0;
+    size_t out_sz = 0;
     if (encode) out = ricecodec_compress(in, in_sz, &out_sz);
     else out = ricecodec_decompress(in, in_sz, &out_sz);
