+++ playground/ricecodec/ricecodec.h
@@ -8,12 +8,14 @@
 #ifndef RICECODEC_H
 #define RICECODEC_H
 
+#include <stdlib.h>
+
 unsigned char* ricecodec_compress(unsigned char* in,
-                                  unsigned int   in_size,
-                                  unsigned int*  out_size);
+                                  size_t         in_size,
+                                  size_t*        out_size);
 unsigned char* ricecodec_decompress(unsigned char* in,
-                                    unsigned int   in_size,
-                                    unsigned int*  out_size);
+                                    size_t         in_size,
+                                    size_t*        out_size);
 
 #endif /* RICECODEC_H */
